Work around 'gperf' not building when WITHOUT_TOOLCHAIN is set
gperf is needed to build gcc for systems that build with gcc (such as
mips). There's a bug where setting WITHOUT_TOOLCHAIN src.conf knob
causes gperf to not get built during the bootstrap compiler phase thus
causing a build break. The in-tree version of gcc is locked at 4.2 and
that's unlikely to change. Work around the problem by manually running
gperf to generate 'cfns.h' and committing the generated header file
along with a cc/cc1plus/Makefile change.