build dnsmasq-2.80
Build dnsmasq-2.80 along with the base system. Install sample
configuration file that starts dnsmasq in dhcp-server only mode.

The included rc script is based on the one found in the ports build. The
variable names have changes but otherwise is still the same.

dnsmasq can be started by enabling it rc.conf. Additional flags may be
supplied as well.

	local_dnsmasq_enable="YES"
	local_dnsmasq_flags="..."