implement basic 'savecfg' mechanism
Modify 'rc' such that on startup, if the trigger file '/etc/initcfg'
exists, run /etc/rc.initcfg just after 'diskless'. This causes a
configuration archive (/cfg/startup-config.tar) to get extracted in
the root of the filesystem if it exists. Any files that would be
overwritten are renamed containing a suffix '.savecfg'. The /cfg
partition must be listed in /etc/fstab. It is up to the appliance
image building script to set this up. This script also will rename
network interfaces according to a 1:1 mapping of <new:old> names
specified in a file called /etc/iface.map if it exists. This file does
not exist by default. It is created by the appliance building script.

The configuration archive is created & written out when root user
invokes the 'savecfg' utility. This is a shell script that upgrades
/cfg to read-write, creates the archive using files listed in
'/etc/savecfg.conf', and then downgrades the /cfg mount.

The reference appliance building script is updated to activate this
feature. It also writes out a '/etc/defaults/vendor.conf' config that
enables sshd & IPv6/SLAAC network configuration on 'gi1'.

A reference DHCPv4-only dnsmasq configuration is included.

The default sshd config is modified to allow root user to log using
keys only. Password login is disabled all together.