Adjust rpi.dts 'memory' range for better FDT_DTB_STATIC behavior
Without this change using an embedded rpi.dts yields only 128MB of
real memory because the firmware never gets a chance to touch that
up. Bump this up to a value that also works with the default
amount of memory allocated to the video core.

With this change, FreeBSD boots with the following parameters:

    real memory  = 469757952 (447 MB)
    avail memory = 428773376 (408 MB)