ds1672 RTC: prevent panic on startup, add FDT support
This driver was causing a panic on startup because it called the
iicbus transfer function before timecounters were enabled. The
iicbus implementation on the RPI does a timed sleep which caused
the panic. Rework driver to lazy init the configuration register.

As a debug aid, RTC 'seconds' register may be printed with the
following sysctl.

  root@:~ # sysctl dev.ds1672_rtc.0.seconds
  dev.ds1672_rtc.0.seconds: 1510529398

i2c bus 1 is routed out to pins 3 & 5 (GPIO 2/3) on the Raspberry Pi
Model B+ 40 pin header.