Allow loader.efi to load from embedded memory disk
The FreeBSD Foundation article mentions embedding a MD image
in loader.efi. This article appears to be a couple years old
and some time between then and now MD image embedding broke.

  https://www.freebsdfoundation.org/freebsd-uefi-secure-boot/

One problem is that the EFI loader's conf.c does not include
the md_dev in its devsw[] pointer list if MD_IMAGE_SIZE is
defined. The second problem is that the EFI loader's main.c
doesn't try to probe the md device as "currdev". The actual
image itself was being correctly included in loader.efi.

Fix those two problems, which allows the EFI loader to load
from the embedded memory disk when MD_IMAGE_SIZE is defined.