build rpi_mini appliance as a Flat Image Tree
This requires u-boot-tools installed on the build host so mkimage
can built a FIT. This also fixes gpio pins 4,5,6 in the dtsi file
to be inputs instead of output / alt function. This is needed to
prevent driver contention when used with the opto22 relay rack.

The FIT contains configurations for a generic RPI and RPI with the
purple board. Booting this FIT requires a U-Boot compiled with
CONFIG_SYS_BOOTM_LEN of about 64MB. github.com/northwoodlogic/u-boot
is known to work.

Booting may be done like this:

    U-Boot> dhcp
    U-Boot> tftp 0x10000000 rpi-mini.itb
    U-Boot> bootm 0x10000000#purple