Add opto22 rack polling daemon
'opto22rackd' uses the opto22 rack driver and polls at a
configurable rate. Values are sent via IPv6 link-scoped multicast
address and port combo [FF02::3]:1900.

Enabling via rc.conf is as follows: This uses interface 'ue0', sends
on a 500mS interval, and sets the message id as '42'.

    opto22rackd_enable="YES"
    opto22rackd_flags="-i ue0 -p 500 -m 42"

Sample golang based (not checked in) multicat receiver output:

  2020/01/25 18:23:18 16 bytes from [fe80::ba27:ebff:fe70:dbea%em0]:47771
  CAN ID: 42, DLC: 4, DATA: 0x00000000

The data structure is modeled after SocketCAN, but is not in anyway
otherwise related.